Br refactor temporal dataset

+++ tgrag/encoders/norm_encoding.py
@@ -7,6 +7,7 @@ from tgrag.encoders.encoder import Encoder
 
 class NormEncoder(Encoder):
     def __call__(self, input: np.ndarray) -> Tensor:
+        input = input.astype(np.float32)
         if input.ndim == 1:
             input = input.reshape(-1, 1)
 
@@ -14,4 +15,4 @@ class NormEncoder(Encoder):
         std = input.std(axis=0, keepdims=True) + 1e-8
 
         normalized = (input - mean) / std
-        return torch.tensor(normalized)
+        return torch.tensor(normalized, dtype=torch.float32)
